Frontend: move most of 'resolution.common.jvm' into 'frontend.common.jvm'
In a subsequent commit, `VirtualFileFinder` will be moved from
'frontend.java' to 'frontend.common.jvm' to remove dependencies of JVM
backend on the old frontend. Apparently, `VirtualFileFinder` depends on
Java elements structure and implementation through
`VirtualFileBoundJavaClass`, so it makes sense to have all of this in
the same module.

Also, 'resolution.common.jvm' seems to be a weird module because it has
the same dependencies as 'frontend.common.jvm' (except
'resolution.common') and almost all its usages also depend on
'frontend.common.jvm'. The only meaningful remaining part is the class
`JvmTypeSpecificityComparator` which is an implementation of a service
from 'resolution.common', and thus cannot be declared in
'frontend.common.jvm'.