[FIR] Add missing test data for ResolveByStubTestGenerated.testAll()
The missing output causes the IDE test to fail.

The test data is reused by the compiler tests (the relevant one was
added in c774aeef3ef42a9cd573281eab0d3beddee97e12, and in the IDE
tests). The long-term plan is to move all stub tests to the Kotlin
repository.