[FIR] Java: rewrite arguments mapping to Java approach for binary sessions
All logic around Java annotations is based on `JavaAnnotation`,
so parameter search should be consistent with this.
Specifically, `lookupTag` is based on `JavaAnnotation#classId`,
which uses Java resolution, so it is supposed to represent `JavaClass`
from`JavaAnnotation#resolve`, but `lookupTag#toSymbol` might result
in a different class.
The main difference is how search works for `JavaAnnotation#resolve`
and for the `lookupTag#toSymbol`. The first one is based on
`JavaClassFinder` directly, while the second uses symbol providers.
Such `FirSession` adds more restrictions and might result in an
incorrect class (or its absence).
And this inconsistency is a blocker for KT-76489. Right now in the
Analysis API mode binary modules have dependencies, which is incorrect
as only resolvable modules may have them. And without such dependencies,
the compiler providers and the Analysis API start to behave differently.
In the compiler, there is a classpath-based symbol provider, which
effectively has all library dependencies based on a use-site module.
So, it will be possible to search for classes from one library in
another one as long as the use-site module has them both in the
classpath.
On the other hand, the Analysis API has granular library modules split,
so either all binary library modules have to depend on all others or
all binary modules have to have proper direct dependencies.

To fix those inconsistencies and to not depend on binary library
dependencies, a more independent and general mechanism must be used.
And Java resolution stands as such the mechanism.

^KT-76540 Fixed

(cherry picked from commit 05b8d1d78cf4170136e91c589dbf9961eea4b4ba)