[FIR] Support negative types in contracts
Not sure if we want to support enums and booleans in contracts, though.
While conditions of the form `param == false` or `param == EnumEntry`
result into an `ERROR_IN_CONTRACT_DESCRIPTION`, it's already possible
to write `param !is Type`.

If we decide to change it, the relevant place is `visitEqualityOperatorCall`
in `ConeEffectExtractor`.

^KT-76635